리셋 방식 수정 / 롬 추가

+++ AppleIIEmulator/gui/raylib/include/raygui.h
@@ -3202,8 +3202,8 @@ Vector2 GuiGrid(Rectangle bounds, const char *text, float spacing, int subdivs)
         if (CheckCollisionPointRec(mousePoint, bounds))
         {
             // NOTE: Cell values must be rounded to int
-            currentCell.x = (int)((mousePoint.x - bounds.x)/spacing);
-            currentCell.y = (int)((mousePoint.y - bounds.y)/spacing);
+            currentCell.x = ((mousePoint.x - bounds.x)/spacing);
+            currentCell.y = ((mousePoint.y - bounds.y)/spacing);
         }
     }
     //--------------------------------------------------------------------
